+++ regions.h
@@ -11,9 +11,9 @@
 extern void regions_add (const char *reg_str);
 extern void regions_make_chregs(unsigned chrom_did_i);
 extern void regions_transform_negative_to_positive_complement(void);
-extern bool regions_get_ra_intersection (uint32_t chrom_node_index, uint32_t min_pos, uint32_t max_pos, char *intersection_one_ra);
+extern bool regions_get_ra_intersection (uint32_t chrom_node_index, int64_t min_pos, int64_t max_pos, char *intersection_one_ra);
 extern unsigned regions_max_num_chregs(void);
 extern void regions_display(const char *title);
-extern bool regions_is_site_included (uint32_t chrom_word_index, uint32_t pos);
+extern bool regions_is_site_included (uint32_t chrom_word_index, int64_t pos);
 
 #endif
